fixed some build warnings

+++ src/PdfToExcel.h
@@ -23,7 +23,7 @@ private:
     std::vector<processor::types::HorizontalLine*> CollectHorizontalLines(size_t nPage);
     const processor::types::HorizontalLine* DoesTextHaveAUnderline(const processor::pdf::types::TextAtLocation* text, const std::vector<processor::types::HorizontalLine*>& lines, const struct Rectangle* mb);
 
-    bool ProcessRows(const processor::pdf::types::TextAtLocation& text_data, int& iLastCol);
+    bool ProcessRows(const processor::pdf::types::TextAtLocation& text_data, size_t& iLastCol);
     void SetupColumns(std::vector<const processor::pdf::types::TextAtLocation*>& header_items);
     void TraceBackForTrade(const std::vector<processor::pdf::types::TextAtLocation>& texts, size_t header_index);
     void ProcessAllTextNodes(const std::vector<processor::pdf::types::TextAtLocation>& all_texts);
